rename grid2 -> grid
