+++ src/views/grid.js
@@ -1,16 +1,24 @@
 const html = require('choo/html')
 const css = require('sheetify')
 const keyboard = require('keyboardjs')
+const HyperList = require('hyperlist-component')
 const offset = require('mouse-event-offset')
 
 const setCursor = require('../util').setCursor
 const contextMenu = require('../components/context-menu.js')
-
 const prefix = css('./grid.css')
+
+const opts = {
+  height: 500,
+  itemHeight: 34,
+  eachItem: tableRow,
+  getTotal: (state) => state.rows.length
+}
+const hyperList = new HyperList('tbody', opts)
 
 module.exports = function grid (state, emit) {
   const { fields, rows } = state.store.activeSheet
-  const selectedCell = state.ui.selectedCell
+  const selectedCell = Object.assign({}, state.ui.selectedCell)
 
   const moveCellUp = moveCell.bind(this, 'up')
   const moveCellDown = moveCell.bind(this, 'down')
@@ -18,44 +26,22 @@ module.exports = function grid (state, emit) {
   const moveCellRight = moveCell.bind(this, 'right')
   const noop = () => {}
 
+  const tbody = hyperList.render({ fields, rows, selectedCell })
+  tbody.onclick = onClickCell
+  tbody.ondblclick = onDblClickCell
+  tbody.addEventListener('blur', onBlurCell, true)
+
   return html`
-    <div class=${prefix}>
-      <table class="table is-bordered is-striped is-narrow"
-             onload=${onload} onunload=${onunload}>
-        <thead oncontextmenu=${showHeaderMenu}>
-          <tr>
-            ${fields.map(tableHeader)}
-          </tr>
+    <div class=${prefix} onload=${onLoad} onunload=${onUnload}>
+      <table class="table is-bordered is-striped is-narrow">
+        <thead oncontextmenu=${onHeaderMenu}>
+          <tr>${fields.map(tableHeader)}</tr>
         </thead>
-        <tbody onclick=${clickCell} ondblclick=${dblClickCell}>
-          ${rows.map(tableRow)}
-        </tbody>
+        ${tbody}
       </table>
       ${headerMenu(state.ui.headerMenu)}
     </div>
   `
-
-  function clickCell (evt) {
-    const el = evt.target
-    const rowIndex = +el.dataset.rowIndex
-    const columnIndex = +el.dataset.columnIndex
-    const isSelected = el.classList.contains('selected')
-    if (!isSelected) {
-      const payload = { rowIndex, columnIndex, editing: false }
-      emit('ui:selectCell', payload)
-    }
-  }
-
-  function dblClickCell (evt) {
-    const el = evt.target
-    const rowIndex = +el.dataset.rowIndex
-    const columnIndex = +el.dataset.columnIndex
-    const isEditing = el.classList.contains('editing')
-    if (!isEditing) {
-      const payload = { rowIndex, columnIndex, editing: true }
-      emit('ui:selectCell', payload)
-    }
-  }
 
   function headerMenu (headerMenuState) {
     const items = [
@@ -67,87 +53,7 @@ module.exports = function grid (state, emit) {
       : ''
   }
 
-  function onload (el) {
-    keyboard.bind('up', moveCellUp)
-    keyboard.bind('down', moveCellDown)
-    keyboard.bind('left', moveCellLeft)
-    keyboard.bind('right', moveCellRight)
-    keyboard.bind('tab', moveCellRight)
-    keyboard.bind('shift + tab', moveCellLeft)
-    keyboard.bind('shift + enter', noop) // differentiate from just enter
-    keyboard.bind('enter', enter)
-  }
-
-  function onunload (el) {
-    keyboard.unbind('up', moveCellUp)
-    keyboard.unbind('down', moveCellDown)
-    keyboard.unbind('left', moveCellLeft)
-    keyboard.unbind('right', moveCellRight)
-    keyboard.unbind('tab', moveCellRight)
-    keyboard.unbind('shift + tab', moveCellLeft)
-    keyboard.unbind('shift + enter', noop)
-    keyboard.unbind('enter', enter)
-  }
-
-  function tableHeader (field) {
-    return html`
-      <th>
-        ${field.name}
-      </th>
-    `
-  }
-
-  function tableRow (row, rowIndex) {
-    return html`
-      <tr>
-        ${fields.map((field, columnIndex) => {
-          const value = row[field.name] || ''
-
-          if (isSelectedCell(rowIndex, columnIndex)) {
-            if (selectedCell.editing) {
-              return tableCellEditing(value)
-            } else {
-              return tableCellSelected(value, rowIndex, columnIndex)
-            }
-          } else {
-            return tableCellDeselected(value, rowIndex, columnIndex)
-          }
-        })}
-      </tr>
-    `
-  }
-
-  function tableCellEditing (value) {
-    return html`
-      <td class="selected editing"
-          contenteditable="true"
-          onblur=${onBlurCell}
-          onload=${setCursorInSelectedCell}>
-        ${value}
-      </td>
-    `
-  }
-
-  function tableCellSelected (value, rowIndex, columnIndex) {
-    return html`
-      <td class="selected"
-          data-row-index=${rowIndex}
-          data-column-index=${columnIndex}>
-        ${value}
-      </td>
-    `
-  }
-
-  function tableCellDeselected (value, rowIndex, columnIndex) {
-    return html`
-      <td data-row-index=${rowIndex}
-          data-column-index=${columnIndex}>
-        ${value}
-      </td>
-    `
-  }
-
-  function showHeaderMenu (evt) {
+  function onHeaderMenu (evt) {
     const parentEl = evt.target.parentNode
     const [x, y] = offset(evt, parentEl)
     emit('ui:headerMenu', { x, y, visible: true })
@@ -158,11 +64,27 @@ module.exports = function grid (state, emit) {
     emit('ui:headerMenu', { visible: false })
   }
 
-  function setCursorInSelectedCell () {
-    // Can't use the `el` arg passed because of dom morphing.
-    // It's probably a bug with nanomorph, technically
-    const el = document.querySelector('.selected')
-    setCursor(el)
+  function onClickCell (evt) {
+    const el = evt.target
+    const rowIndex = +el.dataset.rowIndex
+    const columnIndex = +el.dataset.columnIndex
+    const isSelected = el.classList.contains('selected')
+
+    if (!isSelected) {
+      const payload = { rowIndex, columnIndex, editing: false }
+      emit('ui:selectCell', payload)
+    }
+  }
+
+  function onDblClickCell (evt) {
+    const el = evt.target
+    const rowIndex = +el.dataset.rowIndex
+    const columnIndex = +el.dataset.columnIndex
+    const isEditing = el.classList.contains('editing')
+    if (!isEditing) {
+      const payload = { rowIndex, columnIndex, editing: true }
+      emit('ui:selectCell', payload)
+    }
   }
 
   function onBlurCell (evt) {
@@ -174,9 +96,26 @@ module.exports = function grid (state, emit) {
     }
   }
 
-  function isSelectedCell (rowIndex, columnIndex) {
-    return state.ui.selectedCell.rowIndex === rowIndex &&
-      state.ui.selectedCell.columnIndex === columnIndex
+  function onLoad (el) {
+    keyboard.bind('up', moveCellUp)
+    keyboard.bind('down', moveCellDown)
+    keyboard.bind('left', moveCellLeft)
+    keyboard.bind('right', moveCellRight)
+    keyboard.bind('tab', moveCellRight)
+    keyboard.bind('shift + tab', moveCellLeft)
+    keyboard.bind('shift + enter', noop) // differentiate from just enter
+    keyboard.bind('enter', enter)
+  }
+
+  function onUnload (el) {
+    keyboard.unbind('up', moveCellUp)
+    keyboard.unbind('down', moveCellDown)
+    keyboard.unbind('left', moveCellLeft)
+    keyboard.unbind('right', moveCellRight)
+    keyboard.unbind('tab', moveCellRight)
+    keyboard.unbind('shift + tab', moveCellLeft)
+    keyboard.unbind('shift + enter', noop)
+    keyboard.unbind('enter', enter)
   }
 
   function moveCell (direction, evt) {
@@ -229,4 +168,76 @@ module.exports = function grid (state, emit) {
       emit('store:update', { rowIndex, updates })
     }
   }
+}
+
+function tableHeader (field) {
+  return html`
+    <th>
+      ${field.name}
+    </th>
+  `
+}
+
+function tableRow (state, rowIndex) {
+  const { fields, rows, selectedCell } = state
+  const row = rows[rowIndex]
+
+  return html`
+    <tr>
+      ${fields.map((field, columnIndex) => {
+        const value = row[field.name] || ''
+
+        if (isSelectedCell(rowIndex, columnIndex)) {
+          if (selectedCell.editing) {
+            return tableCellEditing(value)
+          } else {
+            return tableCellSelected(value, rowIndex, columnIndex)
+          }
+        } else {
+          return tableCellDeselected(value, rowIndex, columnIndex)
+        }
+      })}
+    </tr>
+  `
+
+  function isSelectedCell (rowIndex, columnIndex) {
+    return rowIndex === selectedCell.rowIndex &&
+      columnIndex === selectedCell.columnIndex
+  }
+}
+
+function tableCellEditing (value) {
+  return html`
+    <td class="selected editing"
+        contenteditable="true"
+        onload=${setCursorInSelectedCell}>
+      ${value}
+    </td>
+  `
+}
+
+function tableCellSelected (value, rowIndex, columnIndex) {
+  return html`
+    <td class="selected"
+        data-row-index=${rowIndex}
+        data-column-index=${columnIndex}>
+      ${value}
+    </td>
+  `
+}
+
+function tableCellDeselected (value, rowIndex, columnIndex) {
+  return html`
+    <td data-row-index=${rowIndex}
+        data-column-index=${columnIndex}>
+      ${value}
+    </td>
+  `
+}
+
+function setCursorInSelectedCell () {
+  // Can't use the `el` arg passed because of dom morphing.
+  // It's probably a bug with nanomorph, technically
+  const el = document.querySelector('.selected')
+  setCursor(el)
 }
